Refactor el contexto de idioma: se ha movido el LanguageContext a un nuevo archivo y se ha actualizado su uso en varios componentes.
