+++ components/footer.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useLanguage } from "@/lib/language-context";
+import { useLanguage } from "@/lib/translations/language-context";
 import Link from 'next/link'
 import Image from 'next/image'
 import logo from '../public/SkedyLogo.png'
