+++ components/hero.tsx
@@ -5,7 +5,7 @@ import SupabaseLogo from "./supabase-logo";
 import Video from 'next-video';
 import botDemo from '/videos/bot-demo.mp4';
 import JoinWaitlist from "@/components/waitlist-form";
-import { useLanguage } from "@/lib/language-context";
+import { useLanguage } from "@/lib/translations/language-context";
 
 export default function Header() {
   const { t } = useLanguage();
