+++ components/menu.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useLanguage } from "@/lib/language-context";
+import { useLanguage } from "@/lib/translations/language-context";
 import { useState } from 'react';
 import { usePathname } from 'next/navigation';
 import LanguageSwitcher from './language-switcher';
